requirements.txt
httpx>=0.27.0
python-dateutil>=2.8.2
tqdm>=4.66.0
pytz>=2024.1
